feat: stop exporting flatten_route_config function
Previously, flatten_route_config was used to flatten the route_config object before passing it to create_link_generator, allowing easy visualization of the flattened object. However, the FlatRoutes type now provides this capability, and keeping this function public prevented breaking changes for performance improvements. Starting from version 9, link_generator now runs flatten_route_config internally, so exporting this function is no longer necessary.
