+++ src/mod.ts
@@ -1,5 +1,4 @@
 import { link_generator } from "./link_generator.ts";
-import { flatten_route_config } from "./flatten_route_config.ts";
 import type {
   DefaultParamValue,
   ExtractRouteData,
@@ -14,7 +13,6 @@ export {
   type DefaultParamValue,
   type ExtractRouteData,
   type FlatRoutes,
-  flatten_route_config,
   link_generator,
   type LinkGenerator,
   type Param,
